requirements.txt
asgiref==3.8.1
attrs==25.1.0
boto3==1.37.1
botocore==1.37.1
Django==5.1.6
django-debug-toolbar==5.0.1
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
filetype==1.2.0
fuzzywuzzy==0.18.0
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Levenshtein==0.26.1
mysqlclient==2.2.7
pillow==11.1.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.1
PyYAML==6.0.2
RapidFuzz==3.12.1
referencing==0.36.2
rpds-py==0.23.1
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
